requirements.txt

Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.2.1
Werkzeug==3.1.3
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9
python-docx==1.1.0
reportlab==4.0.8
matplotlib==3.8.2
seaborn==0.13.0
Jinja2==3.1.6
MarkupSafe==2.1.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==6.2.0
python-dotenv==1.0.1
flask-limiter==3.12
